Move ValueInfo and GlobOptBlockData to new files
No code changes in this commit, just reorg.
