+++ lib/Backend/GlobOpt.cpp
@@ -20707,303 +20707,6 @@ GlobOpt::OptimizeIndirUses(IR::IndirOpnd *indirOpnd, IR::Instr * *pInstr, Value 
 }
 
 bool
-ValueInfo::IsGeneric() const
-{
-    return structureKind == ValueStructureKind::Generic;
-}
-
-bool
-ValueInfo::IsIntConstant() const
-{
-    return IsInt() && structureKind == ValueStructureKind::IntConstant;
-}
-
-bool
-ValueInfo::IsInt64Constant() const
-{
-    return IsInt() && structureKind == ValueStructureKind::Int64Constant;
-}
-
-const IntConstantValueInfo *
-ValueInfo::AsIntConstant() const
-{
-    Assert(IsIntConstant());
-    return static_cast<const IntConstantValueInfo *>(this);
-}
-const Int64ConstantValueInfo *
-ValueInfo::AsInt64Constant() const
-{
-    Assert(IsInt64Constant());
-    return static_cast<const Int64ConstantValueInfo *>(this);
-}
-
-bool
-ValueInfo::IsIntRange() const
-{
-    return IsInt() && structureKind == ValueStructureKind::IntRange;
-}
-
-const IntRangeValueInfo *
-ValueInfo::AsIntRange() const
-{
-    Assert(IsIntRange());
-    return static_cast<const IntRangeValueInfo *>(this);
-}
-
-bool
-ValueInfo::IsIntBounded() const
-{
-    const bool isIntBounded = IsLikelyInt() && structureKind == ValueStructureKind::IntBounded;
-
-    // Bounds for definitely int values should have relative bounds, otherwise those values should use one of the other value
-    // infos
-    Assert(!isIntBounded || static_cast<const IntBoundedValueInfo *>(this)->Bounds()->RequiresIntBoundedValueInfo(Type()));
-
-    return isIntBounded;
-}
-
-const IntBoundedValueInfo *
-ValueInfo::AsIntBounded() const
-{
-    Assert(IsIntBounded());
-    return static_cast<const IntBoundedValueInfo *>(this);
-}
-
-bool
-ValueInfo::IsFloatConstant() const
-{
-    return IsFloat() && structureKind == ValueStructureKind::FloatConstant;
-}
-
-FloatConstantValueInfo *
-ValueInfo::AsFloatConstant()
-{
-    Assert(IsFloatConstant());
-    return static_cast<FloatConstantValueInfo *>(this);
-}
-
-const FloatConstantValueInfo *
-ValueInfo::AsFloatConstant() const
-{
-    Assert(IsFloatConstant());
-    return static_cast<const FloatConstantValueInfo *>(this);
-}
-
-bool
-ValueInfo::IsVarConstant() const
-{
-    return structureKind == ValueStructureKind::VarConstant;
-}
-
-VarConstantValueInfo *
-ValueInfo::AsVarConstant()
-{
-    Assert(IsVarConstant());
-    return static_cast<VarConstantValueInfo *>(this);
-}
-
-bool
-ValueInfo::IsJsType() const
-{
-    Assert(!(structureKind == ValueStructureKind::JsType && !IsUninitialized()));
-    return structureKind == ValueStructureKind::JsType;
-}
-
-JsTypeValueInfo *
-ValueInfo::AsJsType()
-{
-    Assert(IsJsType());
-    return static_cast<JsTypeValueInfo *>(this);
-}
-
-const JsTypeValueInfo *
-ValueInfo::AsJsType() const
-{
-    Assert(IsJsType());
-    return static_cast<const JsTypeValueInfo *>(this);
-}
-
-bool
-ValueInfo::IsArrayValueInfo() const
-{
-    return IsAnyOptimizedArray() && structureKind == ValueStructureKind::Array;
-}
-
-const
-ArrayValueInfo *ValueInfo::AsArrayValueInfo() const
-{
-    Assert(IsArrayValueInfo());
-    return static_cast<const ArrayValueInfo *>(this);
-}
-
-ArrayValueInfo *
-ValueInfo::AsArrayValueInfo()
-{
-    Assert(IsArrayValueInfo());
-    return static_cast<ArrayValueInfo *>(this);
-}
-
-ValueInfo *
-ValueInfo::SpecializeToInt32(JitArenaAllocator *const allocator, const bool isForLoopBackEdgeCompensation)
-{
-    // Int specialization in some uncommon loop cases involving dependencies, needs to allow specializing values of arbitrary
-    // types, even values that are definitely not int, to compensate for aggressive assumptions made by a loop prepass. In all
-    // other cases, only values that are likely int may be int-specialized.
-    Assert(IsUninitialized() || IsLikelyInt() || isForLoopBackEdgeCompensation);
-
-    if(IsInt())
-    {
-        return this;
-    }
-
-    if(!IsIntBounded())
-    {
-        ValueInfo *const newValueInfo = CopyWithGenericStructureKind(allocator);
-        newValueInfo->Type() = ValueType::GetInt(true);
-        return newValueInfo;
-    }
-
-    const IntBoundedValueInfo *const boundedValueInfo = AsIntBounded();
-    const IntBounds *const bounds = boundedValueInfo->Bounds();
-    const IntConstantBounds constantBounds = bounds->ConstantBounds();
-    if(bounds->RequiresIntBoundedValueInfo())
-    {
-        IntBoundedValueInfo *const newValueInfo = boundedValueInfo->Copy(allocator);
-        newValueInfo->Type() = constantBounds.GetValueType();
-        return newValueInfo;
-    }
-
-    ValueInfo *const newValueInfo =
-        constantBounds.IsConstant()
-            ? static_cast<ValueInfo *>(IntConstantValueInfo::New(allocator, constantBounds.LowerBound()))
-            : IntRangeValueInfo::New(allocator, constantBounds.LowerBound(), constantBounds.UpperBound(), false);
-    newValueInfo->SetSymStore(GetSymStore());
-    return newValueInfo;
-}
-
-ValueInfo *
-ValueInfo::SpecializeToFloat64(JitArenaAllocator *const allocator)
-{
-    if(IsNumber())
-    {
-        return this;
-    }
-
-    ValueInfo *const newValueInfo = CopyWithGenericStructureKind(allocator);
-
-    // If the value type was likely int, after float-specializing, it's preferable to use Int_Number rather than Float, as the
-    // former is also likely int and allows int specialization later.
-    newValueInfo->Type() = IsLikelyInt() ? Type().ToDefiniteAnyNumber() : Type().ToDefiniteAnyFloat();
-
-    return newValueInfo;
-}
-
-// SIMD_JS
-ValueInfo *
-ValueInfo::SpecializeToSimd128(IRType type, JitArenaAllocator *const allocator)
-{
-    switch (type)
-    {
-    case TySimd128F4:
-        return SpecializeToSimd128F4(allocator);
-    case TySimd128I4:
-        return SpecializeToSimd128I4(allocator);
-    default:
-        Assert(UNREACHED);
-        return nullptr;
-    }
-
-}
-
-ValueInfo *
-ValueInfo::SpecializeToSimd128F4(JitArenaAllocator *const allocator)
-{
-    if (IsSimd128Float32x4())
-    {
-        return this;
-    }
-
-    ValueInfo *const newValueInfo = CopyWithGenericStructureKind(allocator);
-
-    newValueInfo->Type() = ValueType::GetSimd128(ObjectType::Simd128Float32x4);
-
-    return newValueInfo;
-}
-
-ValueInfo *
-ValueInfo::SpecializeToSimd128I4(JitArenaAllocator *const allocator)
-{
-    if (IsSimd128Int32x4())
-    {
-        return this;
-    }
-
-    ValueInfo *const newValueInfo = CopyWithGenericStructureKind(allocator);
-
-    newValueInfo->Type() = ValueType::GetSimd128(ObjectType::Simd128Int32x4);
-
-    return newValueInfo;
-}
-
-bool
-ValueInfo::GetIsShared() const
-{
-    return IsJsType() ? AsJsType()->GetIsShared() : false;
-}
-
-void
-ValueInfo::SetIsShared()
-{
-    if (IsJsType()) AsJsType()->SetIsShared();
-}
-
-ValueInfo *
-ValueInfo::Copy(JitArenaAllocator * allocator)
-{
-    if(IsIntConstant())
-    {
-        return AsIntConstant()->Copy(allocator);
-    }
-    if(IsIntRange())
-    {
-        return AsIntRange()->Copy(allocator);
-    }
-    if(IsIntBounded())
-    {
-        return AsIntBounded()->Copy(allocator);
-    }
-    if(IsFloatConstant())
-    {
-        return AsFloatConstant()->Copy(allocator);
-    }
-    if(IsJsType())
-    {
-        return AsJsType()->Copy(allocator);
-    }
-    if(IsArrayValueInfo())
-    {
-        return AsArrayValueInfo()->Copy(allocator);
-    }
-    return CopyWithGenericStructureKind(allocator);
-}
-
-bool
-ValueInfo::GetIntValMinMax(int *pMin, int *pMax, bool doAggressiveIntTypeSpec)
-{
-    IntConstantBounds intConstantBounds;
-    if (TryGetIntConstantBounds(&intConstantBounds, doAggressiveIntTypeSpec))
-    {
-        *pMin = intConstantBounds.LowerBound();
-        *pMax = intConstantBounds.UpperBound();
-        return true;
-    }
-
-    Assert(!IsInt());
-    Assert(!doAggressiveIntTypeSpec || !IsLikelyInt());
-    return false;
-}
-
-bool
 GlobOpt::IsPREInstrCandidateLoad(Js::OpCode opcode)
 {
     switch (opcode)
@@ -21051,76 +20754,6 @@ GlobOpt::ImplicitCallFlagsAllowOpts(Func *func)
 }
 
 #if DBG_DUMP
-void ValueInfo::Dump()
-{
-    if(!IsJsType()) // The value type is uninitialized for a type value
-    {
-        char typeStr[VALUE_TYPE_MAX_STRING_SIZE];
-        Type().ToString(typeStr);
-        Output::Print(_u("%S"), typeStr);
-    }
-
-    IntConstantBounds intConstantBounds;
-    if(TryGetIntConstantBounds(&intConstantBounds))
-    {
-        if(intConstantBounds.IsConstant())
-        {
-            Output::Print(_u(" constant:%d"), intConstantBounds.LowerBound());
-            return;
-        }
-        Output::Print(_u(" range:%d - %d"), intConstantBounds.LowerBound(), intConstantBounds.UpperBound());
-    }
-    else if(IsFloatConstant())
-    {
-        Output::Print(_u(" constant:%g"), AsFloatConstant()->FloatValue());
-    }
-    else if(IsJsType())
-    {
-        const JITTypeHolder type(AsJsType()->GetJsType());
-        type != nullptr ? Output::Print(_u("type: 0x%p, "), type->GetAddr()) : Output::Print(_u("type: null, "));
-        Output::Print(_u("type Set: "));
-        Js::EquivalentTypeSet* typeSet = AsJsType()->GetJsTypeSet();
-        if (typeSet != nullptr)
-        {
-            uint16 typeCount = typeSet->GetCount();
-            for (uint16 ti = 0; ti < typeCount - 1; ti++)
-            {
-                Output::Print(_u("0x%p, "), typeSet->GetType(ti));
-            }
-            Output::Print(_u("0x%p"), typeSet->GetType(typeCount - 1));
-        }
-        else
-        {
-            Output::Print(_u("null"));
-        }
-    }
-    else if(IsArrayValueInfo())
-    {
-        const ArrayValueInfo *const arrayValueInfo = AsArrayValueInfo();
-        if(arrayValueInfo->HeadSegmentSym())
-        {
-            Output::Print(_u(" seg: "));
-            arrayValueInfo->HeadSegmentSym()->Dump();
-        }
-        if(arrayValueInfo->HeadSegmentLengthSym())
-        {
-            Output::Print(_u(" segLen: "));
-            arrayValueInfo->HeadSegmentLengthSym()->Dump();
-        }
-        if(arrayValueInfo->LengthSym())
-        {
-            Output::Print(_u(" len: "));
-            arrayValueInfo->LengthSym()->Dump();
-        }
-    }
-
-    if (this->GetSymStore())
-    {
-        Output::Print(_u("\t\tsym:"));
-        this->GetSymStore()->Dump();
-    }
-}
-
 void
 GlobOpt::Dump()
 {
