+++ lib/Backend/GlobOptIntBounds.cpp
@@ -156,417 +156,6 @@ void GlobOpt::ArrayUpperBoundCheckHoistInfo::SetLoop(
     this->headSegmentLengthConstantBounds = headSegmentLengthConstantBounds;
 }
 
-bool ValueInfo::HasIntConstantValue(const bool includeLikelyInt) const
-{
-    int32 constantValue;
-    return TryGetIntConstantValue(&constantValue, includeLikelyInt);
-}
-
-bool ValueInfo::TryGetInt64ConstantValue(int64 *const intValueRef, const bool isUnsigned) const
-{
-    Assert(intValueRef);
-    if (TryGetIntConstantValue(intValueRef, isUnsigned))
-    {
-        return true;
-    }
-    else
-    {
-        int32 int32ValueRef;
-        if (TryGetIntConstantValue(&int32ValueRef, false))
-        {
-            if (isUnsigned)
-            {
-                *intValueRef = (uint)int32ValueRef;
-            }
-            else
-            {
-                *intValueRef = int32ValueRef;
-            }
-            return true;
-        }
-    }
-    return false;
-}
-
-bool ValueInfo::TryGetIntConstantValue(int64 *const intValueRef, const bool isUnsigned) const
-{
-    Assert(intValueRef);
-    if (structureKind == ValueStructureKind::Int64Constant)
-    {
-        *intValueRef = AsInt64Constant()->IntValue();
-        return true;
-    }
-    return false;
-}
-
-bool ValueInfo::TryGetIntConstantValue(int32 *const intValueRef, const bool includeLikelyInt) const
-{
-    Assert(intValueRef);
-    if(!(includeLikelyInt ? IsLikelyInt() : IsInt()))
-    {
-        return false;
-    }
-
-    switch(structureKind)
-    {
-        case ValueStructureKind::IntConstant:
-            if(!includeLikelyInt || IsInt())
-            {
-                *intValueRef = AsIntConstant()->IntValue();
-                return true;
-            }
-            break;
-
-        case ValueStructureKind::IntRange:
-            Assert(includeLikelyInt && !IsInt() || !AsIntRange()->IsConstant());
-            break;
-
-        case ValueStructureKind::IntBounded:
-        {
-            const IntConstantBounds bounds(AsIntBounded()->Bounds()->ConstantBounds());
-            if(bounds.IsConstant())
-            {
-                *intValueRef = bounds.LowerBound();
-                return true;
-            }
-            break;
-        }
-    }
-    return false;
-}
-
-bool ValueInfo::TryGetIntConstantLowerBound(int32 *const intConstantBoundRef, const bool includeLikelyInt) const
-{
-    Assert(intConstantBoundRef);
-
-    if(!(includeLikelyInt ? IsLikelyInt() : IsInt()))
-    {
-        return false;
-    }
-
-    switch(structureKind)
-    {
-        case ValueStructureKind::IntConstant:
-            if(!includeLikelyInt || IsInt())
-            {
-                *intConstantBoundRef = AsIntConstant()->IntValue();
-                return true;
-            }
-            break;
-
-        case ValueStructureKind::IntRange:
-            if(!includeLikelyInt || IsInt())
-            {
-                *intConstantBoundRef = AsIntRange()->LowerBound();
-                return true;
-            }
-            break;
-
-        case ValueStructureKind::IntBounded:
-            *intConstantBoundRef = AsIntBounded()->Bounds()->ConstantLowerBound();
-            return true;
-    }
-
-    *intConstantBoundRef = IsTaggedInt() ? Js::Constants::Int31MinValue : IntConstMin;
-    return true;
-}
-
-bool ValueInfo::TryGetIntConstantUpperBound(int32 *const intConstantBoundRef, const bool includeLikelyInt) const
-{
-    Assert(intConstantBoundRef);
-
-    if(!(includeLikelyInt ? IsLikelyInt() : IsInt()))
-    {
-        return false;
-    }
-
-    switch(structureKind)
-    {
-        case ValueStructureKind::IntConstant:
-            if(!includeLikelyInt || IsInt())
-            {
-                *intConstantBoundRef = AsIntConstant()->IntValue();
-                return true;
-            }
-            break;
-
-        case ValueStructureKind::IntRange:
-            if(!includeLikelyInt || IsInt())
-            {
-                *intConstantBoundRef = AsIntRange()->UpperBound();
-                return true;
-            }
-            break;
-
-        case ValueStructureKind::IntBounded:
-            *intConstantBoundRef = AsIntBounded()->Bounds()->ConstantUpperBound();
-            return true;
-    }
-
-    *intConstantBoundRef = IsTaggedInt() ? Js::Constants::Int31MaxValue : IntConstMax;
-    return true;
-}
-
-bool ValueInfo::TryGetIntConstantBounds(IntConstantBounds *const intConstantBoundsRef, const bool includeLikelyInt) const
-{
-    Assert(intConstantBoundsRef);
-
-    if(!(includeLikelyInt ? IsLikelyInt() : IsInt()))
-    {
-        return false;
-    }
-
-    switch(structureKind)
-    {
-        case ValueStructureKind::IntConstant:
-            if(!includeLikelyInt || IsInt())
-            {
-                const int32 intValue = AsIntConstant()->IntValue();
-                *intConstantBoundsRef = IntConstantBounds(intValue, intValue);
-                return true;
-            }
-            break;
-
-        case ValueStructureKind::IntRange:
-            if(!includeLikelyInt || IsInt())
-            {
-                *intConstantBoundsRef = *AsIntRange();
-                return true;
-            }
-            break;
-
-        case ValueStructureKind::IntBounded:
-            *intConstantBoundsRef = AsIntBounded()->Bounds()->ConstantBounds();
-            return true;
-    }
-
-    *intConstantBoundsRef =
-        IsTaggedInt()
-            ? IntConstantBounds(Js::Constants::Int31MinValue, Js::Constants::Int31MaxValue)
-            : IntConstantBounds(INT32_MIN, INT32_MAX);
-    return true;
-}
-
-bool ValueInfo::WasNegativeZeroPreventedByBailout() const
-{
-    if(!IsInt())
-    {
-        return false;
-    }
-
-    switch(structureKind)
-    {
-        case ValueStructureKind::IntRange:
-            return AsIntRange()->WasNegativeZeroPreventedByBailout();
-
-        case ValueStructureKind::IntBounded:
-            return AsIntBounded()->WasNegativeZeroPreventedByBailout();
-    }
-    return false;
-}
-
-bool ValueInfo::IsEqualTo(
-    const Value *const src1Value,
-    const int32 min1,
-    const int32 max1,
-    const Value *const src2Value,
-    const int32 min2,
-    const int32 max2)
-{
-    const bool result =
-        IsEqualTo_NoConverse(src1Value, min1, max1, src2Value, min2, max2) ||
-        IsEqualTo_NoConverse(src2Value, min2, max2, src1Value, min1, max1);
-    Assert(!result || !IsNotEqualTo_NoConverse(src1Value, min1, max1, src2Value, min2, max2));
-    Assert(!result || !IsNotEqualTo_NoConverse(src2Value, min2, max2, src1Value, min1, max1));
-    return result;
-}
-
-bool ValueInfo::IsNotEqualTo(
-    const Value *const src1Value,
-    const int32 min1,
-    const int32 max1,
-    const Value *const src2Value,
-    const int32 min2,
-    const int32 max2)
-{
-    const bool result =
-        IsNotEqualTo_NoConverse(src1Value, min1, max1, src2Value, min2, max2) ||
-        IsNotEqualTo_NoConverse(src2Value, min2, max2, src1Value, min1, max1);
-    Assert(!result || !IsEqualTo_NoConverse(src1Value, min1, max1, src2Value, min2, max2));
-    Assert(!result || !IsEqualTo_NoConverse(src2Value, min2, max2, src1Value, min1, max1));
-    return result;
-}
-
-bool ValueInfo::IsEqualTo_NoConverse(
-    const Value *const src1Value,
-    const int32 min1,
-    const int32 max1,
-    const Value *const src2Value,
-    const int32 min2,
-    const int32 max2)
-{
-    return
-        IsGreaterThanOrEqualTo(src1Value, min1, max1, src2Value, min2, max2) &&
-        IsLessThanOrEqualTo(src1Value, min1, max1, src2Value, min2, max2);
-}
-
-bool ValueInfo::IsNotEqualTo_NoConverse(
-    const Value *const src1Value,
-    const int32 min1,
-    const int32 max1,
-    const Value *const src2Value,
-    const int32 min2,
-    const int32 max2)
-{
-    return
-        IsGreaterThan(src1Value, min1, max1, src2Value, min2, max2) ||
-        IsLessThan(src1Value, min1, max1, src2Value, min2, max2);
-}
-
-bool ValueInfo::IsGreaterThanOrEqualTo(
-    const Value *const src1Value,
-    const int32 min1,
-    const int32 max1,
-    const Value *const src2Value,
-    const int32 min2,
-    const int32 max2)
-{
-    return IsGreaterThanOrEqualTo(src1Value, min1, max1, src2Value, min2, max2, 0);
-}
-
-bool ValueInfo::IsGreaterThan(
-    const Value *const src1Value,
-    const int32 min1,
-    const int32 max1,
-    const Value *const src2Value,
-    const int32 min2,
-    const int32 max2)
-{
-    return IsGreaterThanOrEqualTo(src1Value, min1, max1, src2Value, min2, max2, 1);
-}
-
-bool ValueInfo::IsLessThanOrEqualTo(
-    const Value *const src1Value,
-    const int32 min1,
-    const int32 max1,
-    const Value *const src2Value,
-    const int32 min2,
-    const int32 max2)
-{
-    return IsLessThanOrEqualTo(src1Value, min1, max1, src2Value, min2, max2, 0);
-}
-
-bool ValueInfo::IsLessThan(
-    const Value *const src1Value,
-    const int32 min1,
-    const int32 max1,
-    const Value *const src2Value,
-    const int32 min2,
-    const int32 max2)
-{
-    return IsLessThanOrEqualTo(src1Value, min1, max1, src2Value, min2, max2, -1);
-}
-
-bool ValueInfo::IsGreaterThanOrEqualTo(
-    const Value *const src1Value,
-    const int32 min1,
-    const int32 max1,
-    const Value *const src2Value,
-    const int32 min2,
-    const int32 max2,
-    const int src2Offset)
-{
-    return
-        IsGreaterThanOrEqualTo_NoConverse(src1Value, min1, max1, src2Value, min2, max2, src2Offset) ||
-        src2Offset == IntConstMin ||
-        IsLessThanOrEqualTo_NoConverse(src2Value, min2, max2, src1Value, min1, max1, -src2Offset);
-}
-
-bool ValueInfo::IsLessThanOrEqualTo(
-    const Value *const src1Value,
-    const int32 min1,
-    const int32 max1,
-    const Value *const src2Value,
-    const int32 min2,
-    const int32 max2,
-    const int src2Offset)
-{
-    return
-        IsLessThanOrEqualTo_NoConverse(src1Value, min1, max1, src2Value, min2, max2, src2Offset) ||
-        (
-            src2Offset != IntConstMin &&
-            IsGreaterThanOrEqualTo_NoConverse(src2Value, min2, max2, src1Value, min1, max1, -src2Offset)
-        );
-}
-
-bool ValueInfo::IsGreaterThanOrEqualTo_NoConverse(
-    const Value *const src1Value,
-    const int32 min1,
-    const int32 max1,
-    const Value *const src2Value,
-    const int32 min2,
-    const int32 max2,
-    const int src2Offset)
-{
-    Assert(src1Value || min1 == max1);
-    Assert(!src1Value || src1Value->GetValueInfo()->IsLikelyInt());
-    Assert(src2Value || min2 == max2);
-    Assert(!src2Value || src2Value->GetValueInfo()->IsLikelyInt());
-
-    if(src1Value)
-    {
-        if(src2Value && src1Value->GetValueNumber() == src2Value->GetValueNumber())
-        {
-            return src2Offset <= 0;
-        }
-
-        ValueInfo *const src1ValueInfo = src1Value->GetValueInfo();
-        if(src1ValueInfo->structureKind == ValueStructureKind::IntBounded)
-        {
-            const IntBounds *const bounds = src1ValueInfo->AsIntBounded()->Bounds();
-            return
-                src2Value
-                    ? bounds->IsGreaterThanOrEqualTo(src2Value, src2Offset)
-                    : bounds->IsGreaterThanOrEqualTo(min2, src2Offset);
-        }
-    }
-    return IntBounds::IsGreaterThanOrEqualTo(min1, max2, src2Offset);
-}
-
-bool ValueInfo::IsLessThanOrEqualTo_NoConverse(
-    const Value *const src1Value,
-    const int32 min1,
-    const int32 max1,
-    const Value *const src2Value,
-    const int32 min2,
-    const int32 max2,
-    const int src2Offset)
-{
-    Assert(src1Value || min1 == max1);
-    Assert(!src1Value || src1Value->GetValueInfo()->IsLikelyInt());
-    Assert(src2Value || min2 == max2);
-    Assert(!src2Value || src2Value->GetValueInfo()->IsLikelyInt());
-
-    if(src1Value)
-    {
-        if(src2Value && src1Value->GetValueNumber() == src2Value->GetValueNumber())
-        {
-            return src2Offset >= 0;
-        }
-
-        ValueInfo *const src1ValueInfo = src1Value->GetValueInfo();
-        if(src1ValueInfo->structureKind == ValueStructureKind::IntBounded)
-        {
-            const IntBounds *const bounds = src1ValueInfo->AsIntBounded()->Bounds();
-            return
-                src2Value
-                    ? bounds->IsLessThanOrEqualTo(src2Value, src2Offset)
-                    : bounds->IsLessThanOrEqualTo(min2, src2Offset);
-        }
-    }
-    return IntBounds::IsLessThanOrEqualTo(max1, min2, src2Offset);
-}
-
 void GlobOpt::UpdateIntBoundsForEqualBranch(
     Value *const src1Value,
     Value *const src2Value,
